Move async event wait helper to utils
